requirements.txt
pyTelegramBotAPI==4.6.0
python-dotenv==0.20.0
loguru~=0.6.0
requests==2.28.1
peewee==3.15.1